+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
+- New feature ([#1247](https://github.com/open-telemetry/weaver/issues/1247), [#1248](https://github.com/open-telemetry/weaver/issues/1248)) - `.weaver.toml` config file for live-check finding overrides and filters. Use `--config` or auto-discovery from CWD. ([#1256](https://github.com/open-telemetry/weaver/pull/1256) by @jerbly)
 - Add support for git references in the registry url ([#182](https://github.com/open-telemetry/weaver/issues/182) by @sebasnabas)
 - New feature ([#1153](https://github.com/open-telemetry/weaver/issues/1153)) - Live-check now has a `/health` endpoint that can be used in long-running scenarios to confirm readiness and liveness of the live-check server. ([#1193](https://github.com/open-telemetry/weaver/pull/1193) by @jerbly)
 - New feature ([#1100](https://github.com/open-telemetry/weaver/issues/1100)) - Set `--output=http` to have live-check send its report as the response to `/stop`. ([#1193](https://github.com/open-telemetry/weaver/pull/1193) by @jerbly)